// src/app/Model/chaseParams.model.ts
export class chaseParams {
        

    auto_renewal_date: string;
    autorenew_id: string;
    emailSentList: any;
        email_date: string;
        email_read_at: string;
        email_sent: string;
        event: string;
      
    lastNotification: string;
    payer: string;
    quoteList: any;
       
        bp_link_id: string;
        comments: string;
        contract_expiry_date: string;
        dealer_partner_name: string;
        expiring_contractnumber: string;
        expiring_ponumber: string;
        is_edited: string;
        is_eligible: string;
        is_promotion_applied: string;
        mea_id: string;
        owcc_id: string;
        parent_qrn: string;
        poStatus: string;
        po_amount: string;
        po_document: string;
        po_expiry_date: string;
        po_issue_date:string;
        po_number: string;
        quoteSubStatusName: string;
        quote_created_date: string;
        quote_end_date: string;
        quote_id: string
        quote_start_date:string;
        quote_status: string;
        quote_sub_status: string;
        reason: string;
        shipto: string;
        total_contract_value: string;
    
    tier: string;
    tier1: string;
    tier2: string;
    upcomingSentList: any;
   t1soldto:any;
       
       
}